Moved to a flat structure - folders just complicated things
AuraAPIActionsService has been tidied up

aura api token is now managed so that tokens are only fetched when needed - first time or expired

examples updated
