+++ DevExGoals.md
@@ -1,0 +1,90 @@
+This decribes the Dev Ex goals for this project.  I acknowledge that this is a work in progress and right now these are aspirational. 
+
+
+
+# 🧭 Go API Wrapper Developer Experience Checklist
+This checklist is intended to help maintain a **great developer experience (DX)** for Go API wrappers.  
+Use it during reviews or before releases to ensure the library feels idiomatic, intuitive, and consistent.
+
+---
+
+## 1️⃣ Setup and Discovery
+- [ ] **Single entrypoint:** There’s one obvious way to create a client (`NewClient(...)`).
+- [ ] **Minimal friction:** Developers can connect and make their first call in under 3 lines.
+- [ ] **Sensible defaults:** Works without manual config for base URL, timeouts, etc.
+- [ ] **Functional options:** Configurable via `WithX()` options — no massive structs.
+- [ ] **Clear naming:** Functions and types use domain language, not internal API terms.
+
+---
+
+## 2️⃣ API Design and Idiomatic Go
+- [ ] **Idiomatic verbs:** Methods use verbs like `GetUser`, `ListUsers`, `CreateUser`.
+- [ ] **Context everywhere:** Each request accepts a `context.Context`.
+- [ ] **Typed responses:** Returns structs, not `interface{}` or raw maps.
+- [ ] **Simple returns:** Follows `(result, error)` — no custom wrappers unless justified.
+- [ ] **No panics:** Library never panics under normal use.
+
+---
+
+## 3️⃣ Errors and Resilience
+- [ ] **Descriptive errors:** Wrapped with context (e.g. `return fmt.Errorf("get user: %w", err)`).
+- [ ] **Typed or sentinel errors:** Allow `errors.Is(err, ErrUnauthorized)` checks.
+- [ ] **Helpful messages:** Enough detail to diagnose issues without dumping HTTP payloads.
+- [ ] **Retries/backoff:** Optional and safe defaults for transient network failures.
+- [ ] **Rate-limit awareness:** Returns retry hints if supported by the API.
+
+---
+
+## 4️⃣ Structs and JSON Models
+- [ ] **Exported structs:** Public-facing data types are exported.
+- [ ] **Proper JSON tags:** Field names match documented API (`json:"user_id"`).
+- [ ] **Optional fields:** Use `omitempty` for nullable or optional values.
+- [ ] **Stable contracts:** Avoid renaming/removing fields across versions.
+- [ ] **Minimal magic:** Prefer standard `json.Unmarshal` over custom logic unless needed.
+
+---
+
+## 5️⃣ Developer Feedback and Debugging
+- [ ] **Debug mode:** Optional logging for requests and responses.
+- [ ] **Clear error boundaries:** Distinguish between network and API errors.
+- [ ] **Helpful error data:** Includes status code, endpoint, and relevant message.
+- [ ] **Instrumentation hooks:** Optional metrics or event callbacks for observability.
+
+---
+
+## 6️⃣ Testing and Mockability
+- [ ] **Interface-based:** Core client implements a simple `API` interface.
+- [ ] **Mocks or fakes:** Easy to stub out for unit tests.
+- [ ] **No globals:** Avoid shared mutable state or side effects.
+- [ ] **Custom HTTP client:** Supports user-supplied `*http.Client` for testing.
+- [ ] **Robust tests:** Unit tests use `httptest.Server` to simulate API behavior.
+
+---
+
+## 7️⃣ Documentation and Examples
+- [ ] **Godoc coverage:** Every exported symbol has a meaningful doc comment.
+- [ ] **Runnable examples:** In `example_test.go` and verified via `go test`.
+- [ ] **README coverage:** Includes setup, authentication, and first call.
+- [ ] **Error-handling examples:** Show checking specific error types.
+- [ ] **Versioning clarity:** Communicates breaking changes or compatibility guarantees.
+
+---
+
+## 8️⃣ Developer Delight (Bonus Points)
+- [ ] Pagination helpers or iterators instead of manual cursor loops.
+- [ ] Convenience one-liners (`client.MustGetUser(id)` for CLI tools).
+- [ ] Consistent timeout and cancellation handling.
+- [ ] Graceful handling of context cancellations.
+- [ ] Helpful `String()` methods for logging.
+
+---
+
+### ✅ Final DX Test
+> Ask a new developer to use the wrapper for 10 minutes *without reading the docs.*  
+> - Could they figure it out from naming and autocomplete alone?  
+> - If yes → you’ve nailed it.
+
+---
+
+**Tip:** Treat the wrapper like a product.  
+The better it *feels* to use, the more developers will trust and adopt it.
