+++ authMgr.go
@@ -1,0 +1,57 @@
+package aura
+
+import (
+	"context"
+	http "net/http"
+	"net/url"
+	"time"
+
+	httpClient "github.com/LackOfMorals/aura-client/internal/httpClient"
+	utils "github.com/LackOfMorals/aura-client/internal/utils"
+)
+
+// Used for authentication with endpoints
+// Stores the auth token for use with the Aura API
+type APIAuth struct {
+	Type   string `json:"token_type"`
+	Token  string `json:"access_token"`
+	Expiry int64  `json:"expires_in"`
+}
+
+// If needed, updates AuthManager token to make a request to the aura api otherwise it does nothing as the current token is still valid
+func (am *authManager) getToken(ctx context.Context, httpClt httpClient.HTTPService) error {
+	var err error
+
+	// See if we have a token.  If this was the first time this function was called, token will be empty.
+	if len(am.Token) > 0 {
+		// We do have a token, is it still valid?
+		if time.Now().Unix() <= am.ExpiresAt-60 {
+			// We are not within 60 seconds of expiring .  Our token is still valid
+			return nil
+		}
+	}
+
+	// To get a token, we use Basic Auth for the Aura API token endpoint
+	auth := "Basic" + " " + utils.Base64Encode(am.Id, am.Secret)
+
+	endpoint := "oauth/token"
+
+	body := url.Values{}
+
+	body.Set("grant_type", "client_credentials")
+
+	newAuraAPIToken, err := makeAuthenticatedRequest[APIAuth](ctx, httpClt, auth, endpoint, http.MethodPost, "application/x-www-form-urlencoded", body.Encode())
+	if err != nil {
+		// Didn't get a token
+		return err
+	}
+
+	// Update the token details
+	am.ObtainedAt = time.Now().Unix()
+	am.Token = newAuraAPIToken.Token
+	am.Type = newAuraAPIToken.Type
+	am.ExpiresAt = time.Now().Unix() + newAuraAPIToken.Expiry
+
+	return nil
+
+}
