+++ examples/updateInstance/main.go
@@ -9,7 +9,6 @@ import (
 	"os"
 
 	"github.com/LackOfMorals/aura-client"
-	"github.com/LackOfMorals/aura-client/resources"
 )
 
 const (
@@ -66,18 +65,12 @@ func main() {
 		os.Exit(1)
 	}
 
-	auraToken, err := myAuraClient.Auth.GetAuthToken(ctx)
-	if err != nil {
-		log.Println("Error getting token: ", err)
-		os.Exit(1)
-	}
-
-	instanceNewCfg := resources.UpdateInstanceData{
+	instanceNewCfg := aura.UpdateInstanceData{
 		Name:   "JGNewName",
 		Memory: "8GB",
 	}
 
-	response, err := myAuraClient.Instances.Update(ctx, auraToken, instanceID, &instanceNewCfg)
+	response, err := myAuraClient.Instances.Update(ctx, instanceID, &instanceNewCfg)
 	if err != nil {
 		log.Println("Error creating instance: ", err)
 		os.Exit(1)
