--- resources/types.go
@@ -1,196 +1,0 @@
-package resources
-
-import (
-	"time"
-
-	httpClient "github.com/LackOfMorals/aura-client/internal/httpClient"
-)
-
-// Core service configuration
-type AuraAPIActionsService struct {
-	BaseURL      string
-	Version      string
-	ClientID     string
-	ClientSecret string
-	Timeout      time.Duration
-
-	Http httpClient.HTTPService
-
-	// Grouped services
-	Auth           *AuthService
-	Tenants        *TenantService
-	Instances      *InstanceService
-	Snapshots      *SnapshotService
-	Cmek           *CmekService
-	GraphAnalytics *GDSSessionService
-}
-
-// The following structs that are used for requests / responses to and from the Aura API
-// they are models for interacting with the Aura API
-
-// Stores the auth token for use with the Aura API
-type AuthAPIToken struct {
-	Type   string `json:"token_type"`
-	Token  string `json:"access_token"`
-	Expiry int64  `json:"expires_in"`
-}
-
-// Tenants
-
-// A list of tenants in your organisation, each with summary data
-type ListTenantsResponse struct {
-	Data []TenantsRepostData `json:"data"`
-}
-
-type TenantsRepostData struct {
-	Id   string `json:"id"`
-	Name string `json:"name"`
-}
-
-// Details of a tenant
-type GetTenantResponse struct {
-	Data TenantRepostData `json:"data"`
-}
-
-type TenantRepostData struct {
-	Id                     string                        `json:"id"`
-	Name                   string                        `json:"name"`
-	InstanceConfigurations []TenantInstanceConfiguration `json:"instance_configurations"`
-}
-
-type TenantInstanceConfiguration struct {
-	CloudProvider string `json:"cloud_provider"`
-	Region        string `json:"region"`
-	RegionName    string `json:"region_name"`
-	Type          string `json:"type"`
-	Memory        string `json:"memory"`
-	Storage       string `json:"storage"`
-	Version       string `json:"version"`
-}
-
-// Instances
-
-// List of instances in a tenant
-type ListInstancesResponse struct {
-	Data []ListInstanceData `json:"data"`
-}
-
-type ListInstanceData struct {
-	Id            string `json:"id"`
-	Name          string `json:"name"`
-	Created       string `json:"created_at"`
-	TenantId      string `json:"tenant_id"`
-	CloudProvider string `json:"cloud_provider"`
-}
-
-type CreateInstanceConfigData struct {
-	Name          string `json:"name"`
-	TenantId      string `json:"tenant_id"`
-	CloudProvider string `json:"cloud_provider"`
-	Region        string `json:"region"`
-	Type          string `json:"type"`
-	Version       string `json:"version"`
-	Memory        string `json:"memory"`
-}
-
-type CreateInstanceResponse struct {
-	Data CreateInstanceData `json:"data"`
-}
-
-type CreateInstanceData struct {
-	Id            string `json:"id"`
-	Name          string `json:"name"`
-	TenantId      string `json:"tenant_id"`
-	CloudProvider string `json:"cloud_provider"`
-	ConnectionUrl string `json:"connection_url"`
-	Region        string `json:"region"`
-	Type          string `json:"type"`
-	Username      string `json:"username"`
-	Password      string `json:"password"`
-}
-
-type UpdateInstanceData struct {
-	Name   string `json:"name"`
-	Memory string `json:"memory"`
-}
-
-type GetInstanceResponse struct {
-	Data GetInstanceData `json:"data"`
-}
-
-type GetInstanceData struct {
-	Id              string  `json:"id"`
-	Name            string  `json:"name"`
-	Status          string  `json:"status"`
-	TenantId        string  `json:"tenant_id"`
-	CloudProvider   string  `json:"cloud_provider"`
-	ConnectionUrl   string  `json:"connection_url"`
-	Region          string  `json:"region"`
-	Type            string  `json:"type"`
-	Memory          string  `json:"memory"`
-	Storage         *string `json:"storage"`
-	CDCEnrichment   string  `json:"cdc_enrichment_mode"`
-	GDSPlugin       bool    `json:"graph_analytics_plugin"`
-	MetricsURL      string  `json:"metrics_integration_url"`
-	Secondaries     int     `json:"secondaries_count"`
-	VectorOptimized bool    `json:"vector_optimized"`
-}
-
-type GetSnapshotsResponse struct {
-	Data []GetSnapshotData `json:"data"`
-}
-
-type GetSnapshotData struct {
-	InstanceId string `json:"instance_id"`
-	SnapshotId string `json:"snapshot_id"`
-	Profile    string `json:"profile"`
-	Status     string `json:"status"`
-	Timestamp  string `json:"timestamp"`
-}
-
-type GetSnapshotResponse struct {
-	Data GetSnapshotData `json:"data"`
-}
-
-type CreateSnapshotResponse struct {
-	Data CreateSnapshotData `json:"data"`
-}
-
-type CreateSnapshotData struct {
-	SnapshotId string `json:"snapshot_id"`
-}
-
-// Customer Managed Encryption Keys
-
-type GetCmeksResponse struct {
-	Data []GetCmeksData `json:"data"`
-}
-
-type GetCmeksData struct {
-	Id       string `json:"id"`
-	Name     string `json:"name"`
-	TenantId string `json:"tenant_id"`
-}
-
-// GDS Sessions
-
-type GetGDSSessionResponse struct {
-	Data []GetGDSSessionData `json:"data"`
-}
-
-type GetGDSSessionData struct {
-	Id            string `json:"id"`
-	Name          string `json:"name"`
-	Memory        string `json:"memory"`
-	InstanceId    string `json:"instance_id"`
-	DatabaseId    string `json:"database_uuid"`
-	Status        string `json:"status"`
-	Create        string `json:"created_at"`
-	Host          string `json:"host"`
-	Expiry        string `json:"expiry_date"`
-	Ttl           string `json:"ttl"`
-	UserId        string `json:"user_id"`
-	TenantId      string `json:"tenant_id"`
-	CloudProvider string `json:"cloud_provider"`
-	Region        string `json:"region"`
-}
