+++ types.go
@@ -1,0 +1,22 @@
+package aura
+
+/*
+// The following structs that are used for requests / responses to and from the Aura API
+// they are models for interacting with the Aura API
+
+// Stores the auth token for use with the Aura API
+type AuthAPIToken struct {
+	Type   string `json:"token_type"`
+	Token  string `json:"access_token"`
+	Expiry int64  `json:"expires_in"`
+}
+
+// Used for authentication with endpoints
+// Stores the auth token for use with the Aura API
+type APIAuth struct {
+	Type   string `json:"token_type"`
+	Token  string `json:"access_token"`
+	Expiry int64  `json:"expires_in"`
+}
+
+*/
